Fixed an issue where the poll was having wrong result after receiving a websocket event.

+++ messaging/GLPPollOperationManager.m
@@ -128,7 +128,7 @@ static GLPPollOperationManager *instance = nil;
 
 - (void)pollUpdatedWithPollRemoteKey:(NSInteger)pollRemoteKey withNewPollData:(GLPPoll *)pollData
 {
-    DDLogDebug(@"GLPPollOperationManager updatePoll notification %@ remote key %ld", pollData, pollRemoteKey);
+    DDLogDebug(@"GLPPollOperationManager updatePoll notification %@ remote key %ld", pollData, (long)pollRemoteKey);
     
     [[NSNotificationCenter defaultCenter] postNotificationName:GLPNOTIFICATION_POLL_VIEW_STATUS_CHANGED object:self userInfo:@{@"poll_remote_key": @(pollRemoteKey), @"kind_of_operation" : @(kPollUpdated), @"poll_updated_data" : pollData}];
 }
